restartmanager: move to daemon/internal
The restartmanager is only used internally by the daemon, and has no external
users. Move it to the daemon/internal package.
