+++ daemon/internal/restartmanager/restartmanager.go
