+++ daemon/internal/restartmanager/restartmanager_test.go
